+++ tests/workers/rollout/test_sglang_async_rollout_w_tools.py
@@ -35,8 +35,8 @@ from utils_sglang import (
 )
 
 from verl import DataProto
-from verl.workers.rollout.sglang_rollout.async_sglang_rollout import AsyncSGLangRollout
-from verl.workers.sharding_manager.fsdp_sglang import FSDPAsyncSGLangShardingManager
+from verl.workers.rollout.sglang_rollout.sglang_rollout import SGLangRollout
+from verl.workers.sharding_manager.fsdp_sglang import FSDPSGLangShardingManager
 
 
 def test_async_sglang_rollout_w_tool():
@@ -78,9 +78,9 @@ def test_async_sglang_rollout_w_tool():
     )
 
     rollout_config = get_rollout_config(max_response_length, max_prompt_length, dtype, tensor_parallel_size)
-    rollout = AsyncSGLangRollout(actor_module=local_model_path, config=rollout_config, tokenizer=tokenizer, model_hf_config=actor_model.config)
+    rollout = SGLangRollout(actor_module=local_model_path, config=rollout_config, tokenizer=tokenizer, model_hf_config=actor_model.config)
 
-    rollout_sharding_manager = FSDPAsyncSGLangShardingManager(
+    rollout_sharding_manager = FSDPSGLangShardingManager(
         module=fsdp_model,
         inference_engine=rollout._engine,
         model_config=actor_model.config,
