+++ verl/trainer/config/ppo_megatron_trainer.yaml
@@ -152,7 +152,7 @@ actor_rollout_ref:
       n: 1
       do_sample: False # default eager for validation
     multi_turn: 
-      enable: False  # should set rollout.name to sglang_async if True
+      enable: False  # should set rollout.name to sglang if True
       max_turns: null  # null for no limit (default max_length // 3)
       tool_config_path: null  # null for no tool
       format: chatml  # chatml, more formats will be supported in the future
